# remove some defer and reduce memory allocation