`pprof` shows `defer` in `get` and `getTableForKey` takes lots of time, so remove them in these two function.

Due to limitation of Go Compiler's escape analysis, the `itr.key = itr.keyBuf[:0]` will cause the whole `blockIterator` escape to heap. https://github.com/golang/go/issues/7921#issuecomment-285855447.

After some refactor, we can avoid this allocation when key length less than 32 bytes. The benchmark result is

Before:
```
BenchmarkPointGet/Seek_1000-12         	   20000	    453132 ns/op	  448000 B/op	    1000 allocs/op
BenchmarkPointGet/Hash_1000-12         	   20000	    391929 ns/op	  448000 B/op	    1000 allocs/op
BenchmarkPointGet/Seek_10000-12        	    2000	   5529729 ns/op	 4480002 B/op	   10000 allocs/op
BenchmarkPointGet/Hash_10000-12        	    2000	   4623209 ns/op	 4480002 B/op	   10000 allocs/op
BenchmarkPointGet/Seek_100000-12       	     100	  87278133 ns/op	44800053 B/op	  100000 allocs/op
BenchmarkPointGet/Hash_100000-12       	     100	  76633040 ns/op	44800053 B/op	  100000 allocs/op
BenchmarkPointGet/Seek_1000000-12      	       5	1420931867 ns/op	448001075 B/op	 1000000 allocs/op
BenchmarkPointGet/Hash_1000000-12      	       5	1210990874 ns/op	448001075 B/op	 1000000 allocs/op
BenchmarkPointGet/Seek_5000000-12      	       1	8763558217 ns/op	2240005376 B/op	 5000001 allocs/op
BenchmarkPointGet/Hash_5000000-12      	       1	7088176634 ns/op	2240005376 B/op	 5000001 allocs/op
BenchmarkPointGet/Seek_10000000-12     	       1	18787675792 ns/op	4480005376 B/op	10000001 allocs/op
BenchmarkPointGet/Hash_10000000-12     	       1	15630159271 ns/op	4480005376 B/op	10000001 allocs/op
BenchmarkPointGet/Seek_15000000-12     	       1	29897498825 ns/op	6720005376 B/op	15000001 allocs/op
BenchmarkPointGet/Hash_15000000-12     	       1	25130976701 ns/op	6720005376 B/op	15000001 allocs/op
```

After:
```
BenchmarkPointGet/Seek_1000-12         	   20000	    374426 ns/op	       0 B/op	       0 allocs/op
BenchmarkPointGet/Hash_1000-12         	   30000	    299675 ns/op	       0 B/op	       0 allocs/op
BenchmarkPointGet/Seek_10000-12        	    2000	   4554433 ns/op	       2 B/op	       0 allocs/op
BenchmarkPointGet/Hash_10000-12        	    2000	   3519133 ns/op	       2 B/op	       0 allocs/op
BenchmarkPointGet/Seek_100000-12       	     100	  63780739 ns/op	      53 B/op	       0 allocs/op
BenchmarkPointGet/Hash_100000-12       	     100	  53004132 ns/op	      53 B/op	       0 allocs/op
BenchmarkPointGet/Seek_1000000-12      	       5	1254546150 ns/op	    1075 B/op	       0 allocs/op
BenchmarkPointGet/Hash_1000000-12      	       5	1063565797 ns/op	    1075 B/op	       0 allocs/op
BenchmarkPointGet/Seek_5000000-12      	       1	8397103129 ns/op	    5376 B/op	       1 allocs/op
BenchmarkPointGet/Hash_5000000-12      	       1	6564546701 ns/op	    5376 B/op	       1 allocs/op
BenchmarkPointGet/Seek_10000000-12     	       1	18007819630 ns/op	    5376 B/op	       1 allocs/op
BenchmarkPointGet/Hash_10000000-12     	       1	14948745975 ns/op	    5376 B/op	       1 allocs/op
BenchmarkPointGet/Seek_15000000-12     	       1	29151195895 ns/op	    5376 B/op	       1 allocs/op
BenchmarkPointGet/Hash_15000000-12     	       1	24251887305 ns/op	    5376 B/op	       1 allocs/op
```