@bobotu 2018-11-14
Result of benchmark seems wired... I'll reopen this PR later.